refactor(editor): implement Atomic Design structure
- Reorganize components into atoms, molecules, organisms, and templates.

- Break down the editor into smaller, reusable components:
  - Atoms: Button, Input, Icon, etc.
  - Molecules: Toolbar, SearchBar, etc.
  - Organisms: EditorHeader, EditorSidebar, etc.
  - Templates: EditorLayout.
- Update imports and exports to reflect the new structure.
- Ensure consistency and reusability across the design system.
