+++ unisala-astro/src/components/newstory/index.tsx
@@ -1,0 +1,8 @@
+import PublishedPostsList from '@/components/newstory/templates/publishedList';
+import VisualAidTemplate from '@/components/newstory/templates/aidPanel';
+import DashboardMetrics from '@/components/newstory/templates/extendedPanel';
+import PreviewModal from '@/components/newstory/templates/storyPreviewModal';
+import VisualAidPanel from '@/components/newstory/templates/visualAidPanel';
+import DraftsList from '@/components/newstory/templates/draftlist';
+
+export { PublishedPostsList, VisualAidTemplate, DashboardMetrics, PreviewModal, VisualAidPanel, DraftsList };
