+++ unisala-astro/src/components/newstory/templates/saveNotes.tsx
@@ -1,7 +1,7 @@
 import { useAstroMutation } from '@/datasource/apollo-client';
 import { USER_SERVICE_GQL } from '@/datasource/servers/types';
 import toast from 'react-hot-toast';
-import { Button } from '../ui/button';
+import { Button } from '../../ui/button';
 import { ADD_NOTES } from '@/datasource/graphql/user';
 import { getCache } from '@/utils/cache';
 import { navigator } from '@/utils/lib/URLupdate';
