+++ unisala-astro/src/pages/new-story/drafts.astro
@@ -1,13 +1,13 @@
 ---
 import MainLayout from '@/layouts/main-layout.astro';
 import { Toast } from '@/components/ui/toast';
-import DraftsList from '@/components/newstory/draftlist';
+import {DraftsList} from '@/components/newstory/index';
 
 ---
 
 <MainLayout title='My Drafts'>
     <Toast client:load />
-    
+
     <div class="container mx-auto px-4 py-8">
         <h1 class="text-3xl font-bold mb-6">My Drafts</h1>
         <DraftsList client:only />
