+++ unisala-astro/src/pages/new-story/index.astro
@@ -1,6 +1,6 @@
 ---
 import MainLayout from '@/layouts/main-layout.astro';
-import StoryForm from '@/components/newstory/storyform';
+import {StoryForm} from '@/components/newstory/templates/storyform';
 import type { PostDraft } from '@/types/post';
 import {Toast} from '@/components/ui/toast';
 
@@ -21,5 +21,5 @@ const initialPostDraft: PostDraft = {
 <MainLayout title='Create Post'>
     <Toast client:load />
     <StoryForm client:only initialPostDraft={initialPostDraft} />
-    
+
 </MainLayout>
