+++ unisala-astro/src/pages/published.astro
@@ -1,7 +1,7 @@
 ---
 import MainLayout from '@/layouts/main-layout.astro';
 import { Toast } from '@/components/ui/toast';
-import PublishedPostsList from '@/components/newstory/publishedList';
+import {PublishedPostsList} from '@/components/newstory/index';
 
 ---
