+++ unisala-ionic/src/datasource/servers/config.ts
@@ -1,4 +1,4 @@
 const config = {
-  NODE_ENV: "PRODUCTION",
+  NODE_ENV: "DEVELOPMENT",
 };
 export default config;
